Change function openviews to menu.js
